[XLS] Introduce PartialInformation to handle combinations of ternary & range information
This gives us a logical unit we can use to handle interactions of ternary & range constraints. For instance, if we want to take the set of all 3-bit values covered by one of `0b0XX`, `0b10X`, and `0b110`, we can only represent this as `0bXXX` in ternary, but we can see that it's exactly the range `[0, 6]`. In particular, this can let us recognize that the only value in the complement of the above set is  `0b111`.

This will be applied in followup changes to make conditional specialization more powerful, and possibly to provide a more powerful forward-propagating query engine.

PiperOrigin-RevId: 723979970