Re-enable compilation DB. Add build-cleaner.
The old way of doing the compilation DB does not work anymore with
bzlmod, so use http://bant.build/ for it now.

Adapt the make-compilation-db and run-clang-tidy-cached
script to work with that.

Also, add a new run-build-cleaner script to be able to fix
BUILD dependencies in the repository.

This prepares it to be used in a presubmit at some point as
requested before:
https://github.com/google/xls/pull/1605#discussion_r1762139333

PiperOrigin-RevId: 725719727