Use canonical BCR names for repositories.
The canonical BCR repository names were aliased to the
previously used names in the WORKSPACE environment.
This change switches most of them to using the canonical
names directly (including the references in the BUILD
files they are used in).

Not all repo_name aliases could be removed in this change
yet as there are some remaining WORKSPACE dependencies
that refer to them with that name. These should be
handled in separate changes.

Issues: https://github.com/google/xls/issues/931
PiperOrigin-RevId: 725799584