Support concat in type_system_v2.
This required some tweaks in conversion logic because it's the first time we've used builtins internally in fabricated type annotations.

Note that we can invoke stuff like `element_count` internally within typechecking, without any typechecking of builtins, because in that case we are just relying on BytecodeEmitter to run the function.

PiperOrigin-RevId: 726537666