Fix GetWordBitsAtEndFuzz
The fuzz test could try to get bits after the last bit of the input. This could sometimes work depending on the total number of bits in the input. Make the test never ask for bits starting after the last bit and make the implementation uniformly error if a bit after the last bit is requested.

PiperOrigin-RevId: 726556005