Prevent assert conditions from being inlined expressions.
This is probably paranoia, but it seems useful to be sure that if an assertion references a signal multiple times the reference should be resolved to the same value. Also, assertion predicates are referenced multiple times, so inline expressions can make assertions long and unreadable.

While we're in here, we refactor the assertion logic a bit so that op overrides cause assertions to be emitted for Verilog. They were previously unconditionally disabled, but now we only disable assertions if we're not using SystemVerilog AND we don't have a format string for assertions.

Ref: google/xls#1933.
PiperOrigin-RevId: 728423377