More reliable handling of fabricated expressions in type_system_v2.
When we create expressions or annotations internally during conversion, including by cloning, the resulting subtrees are not in the original AST and not hit by the normal type-checking flow. So we need to be more careful with the type checking data for these fabricated ASTs, in order to avoid bogus evaluation failures.

This is a prerequisite for slicing, but there are examples that fail before this change without slicing.

PiperOrigin-RevId: 729128056