[xls/mlir] Add a name_hint attribute to SpawnOp, InstantiateEprocOp and InstantiateExternEprocOp
This allows the user to specify effectively an instance name for the spawn. This will be reflected
in the name given to any EprocOps produced from it. The usual uniquing rules apply, so this just
changes the name stem from `callee.getName()` to `name_hint`.

InstantiateEprocOp and InstantiateExternEprocOp also get a `name` attribute that is derived from this
name_hint. Again, it is optional and will only be populated when `name_hint` is given to a SpawnOp.

PiperOrigin-RevId: 729550394