Support slicing in type_system_v2.
The remaining TODO here is to clamp the bounds for a slice like x[0:8] so that e.g. if x is only 4 bits, the 8 will be coerced to 4. It is somewhat questionable whether we want to preserve that behavior, but assuming we do, it requires an integer clamping builtin in order to be done nicely.

PiperOrigin-RevId: 729618511