[XLS] Enhance array simplification for updates of `array()`s
Even when the updated index is not known, we can still hoist the update above the `array()` operation, enabling better analysis. Since this involves "splitting" the update operation, we only enable this at higher `opt_level`s.

Not much effect in most cases, but this appears to improve results for some future augmentations to `--optimize_for_best_case_throughput`.

PiperOrigin-RevId: 729701623