Support tuples with bits[0] elements in Z3.
These are often created as part of ram modeling. Not being able to prove on them was a major inconvienece.

This does not implement any of the other math operations on zero-width types but since most other ones are trivially removed by constant prop anyway this should not be a big deal.

PiperOrigin-RevId: 731346693