[XLS] Fix potential infinite loop in common subexpression elimination
If the CSE pass identified a dead node as equivalent to another node, it would attempt to replace all of its uses - making no change, but reporting a change. This could cause infinite loops when running CSE to fixed-point, as in our standard optimization pipeline.

Normally, dead-code elimination would prevent this; however, DCE does not eliminate dead invokes (among other nodes), relying on other passes (inlining, in this case) to clean things up!

To fix this, we skip analyzing dead nodes for CSE.

PiperOrigin-RevId: 731396829