Add type inference for range expression
For range expression like  u32:0..u32:5, its type is inferred as uN[32][5], equivalent to the array [u32:0, u32:1, u32:2, u32:3, u32:4].
There is a limitation due to constexpr evaluator not able to report overflow, that if end is less than start, the range expression yields a zero sized array. This is a workaround such that an iteration over such range has no effect, which is probably what the user would intend.

PiperOrigin-RevId: 731796197