Calculate z3 mutex lazily
We were calculating the mutual exclusion between ~every pair of channel predicates in a graph. This could be very time consuming especially since we actually only care about the mutex between operations on the same channel. By rewriting this to only invoke the solver when actually required, codegen performance can be improved significantly. Some blocks can be 250x faster to codegen because there are no channels which are interesting to examine the mutual exclusivity of.

Previously the  MassiveNumberOfChannelsNoChanges test with 20 channels would take ~10 seconds. With the changes the test can run with 2000 channels in 2 seconds. The MassiveNumberOfChannels test goes from taking 30 seconds with 20 channels to 8 seconds with 128 channels.

PiperOrigin-RevId: 732163196