[xls][mlir] Emit well-formed XLS source location infos in xls_translate
`xls_translate` previously assigned an ID to every source file
encountered in the MLIR `Location` of input ops and referenced these IDs
in the XLS `pos` attribute of the produced IR nodes, but did not
actually add these source file names to the XLS package, leading to
invalid `pos` attributes.

This patch fixes this by relying on `Package::AddSourceLocation` to both
construct the `SourceLocation` for nodes and automatically registering
the file with the package.

With this, `xls_translate` can also re-use the file name ID
assignment/lookup from XLS proper.