Explicitly represent channel interfaces in the serialized IR. Representation looks like:
  chan_interface in_ch(direction=receive, kind=streaming, ...)

This provides a unified location to specify attributes of channel interfaces. Previously these were spread between the proc interface channel arguments in the `my_proc<...>` signature and the channel declaration. The proc signature now only indicates the type, direction, and name of elements on the proc interface. Although the interface defined in `my_proc<...>` is no longer technically required it is useful for readability.

Also, add a couple of fields to ChannelInterface and fix some latent bugs where channel interface attributes were not properly handled.

PiperOrigin-RevId: 735446323