[XLS] Fix non-determinism in the BDD-enabled conditional specialization pass
Since we passed the list of specializations to the query engine in a hash map, we ended up non-deterministically inserting the specializations in different orders. If the BDD saturated, this could implicitly create variables in the BDD in a different order... and since BDD complexity is sensitive to variable ordering, this could in turn cause saturation to occur at different points in the analysis on different runs.

We fix this by switching to btree containers of specializations, rather than hash maps; this ensures we iterate over the inserted specializations in a consistent order.

PiperOrigin-RevId: 736309140