[XLS] Deduplicate variables used in BddQueryEngine
We make sure BddQueryEngine avoids creating more than one BDD variable for the same bit when it's simple - and in particular, guarantee that if a node hasn't changed, we always retrieve the same variables for any of its bits that are not represented in the (saturated) BDD.

This may have some performance implications, but we're mostly adding it so that the (newly lazy) BDD query engine stays verifiably consistent over time. Without this, we would routinely end up creating new variables for (e.g.) a Param node, giving us a mismatch between two runs through the same function... in particular, the BddQueryEngine::CheckConsistency could fail when `--debug_optimizations` enabled it.

PiperOrigin-RevId: 736850673