[XLS] Split visibility analysis out of ConditionalSpecializationPass
We reuse the lazy-query-engine components to make this simpler to implement, but can't quite manage to make this a lazy analysis, since changes to a selector can influence the cases it controls, and (for now) our handling of ConditionSets requires the topological ordering to stay fixed. Regardless, this is one step in the right direction, and will let us reuse the same analysis code in other contexts.

Next up: resource sharing!

PiperOrigin-RevId: 736979299