Make proc ast-gen signature-driven.
Previously, the random proc AST generator would start with zero
channels, and, every time a ChannelOp was generated, add a new channel
of the appropriate type.

This updates the generator to first pick a random proc "signature" (i.e.
set of input/output channels) and then generate a matching proc body.
Since all channels must be interacted with, the AST generator now tracks
which channels have already been "used", and will generate additional
ChannelOps for all unused channels at the end of the `next` function if
any are left.

Note that this requires the ability to always generate a `ChannelOp` for
any channel, so the decision logic to pick predicates and values for
sends has been update to generate a new literal value if no matching
value can be found in the environment.

With this change, we can now generate procs of a pre-determined
"signature", just like we can for functions. This will make
generating and fuzzing multi-proc ASTs/procs which span child procs
significantly easier down the road.

In combination, these changes significantly increase the number of channels
per proc and channel-ops seen when fuzzing. This exposes a new AST
generator/fuzzer limitation: The combinational proc codegen backend has
stricter requirements for the exclusivity of output channels that are
not guaranteed by the AST generator, leading to many fuzzer crashes.
This has been added as a "known_failure" since it seems to be a
limitation of the AST generator, not the codegen backend.