Remove restriction that FFI return values can only access scalars.
Before, it was only possible to access return values by scalar, either
because `return` in itself is a scalar, or its tuple values are all the
way decomposed, e.g `return.0.1` can be accessed if that was a scalar,
but `return.0`, which is a tuple, could not.

This removes the restriction, so that it is possible to access tuple
values directly which are expected to be in the same bit-order
XLS expects.

Fixes https://github.com/google/xls/issues/1722

PiperOrigin-RevId: 738095773