Implement downcasting for apfloat when exponent width is also possibly reduced.
This change hides the method downcast_fractional_rne as private to apfloat.x and instead exposes the more general downcast_rne, which also allows the exponent width to be reduced.

This change also introduces a helper method exponent_bias, which returns the constant offset used to encode the exponent for a given type (e.g. 127 for types with an 8-bit exponent), and simplifies the methods bias() and unbiased_exponent().

Co-authored-by: Rasmus Larsen <rmlarsen@google.com>
FolderOrigin-RevId: /google/src/cloud/meheff/clean/google3/..