Use or-tools from Bazel Central Registry.
Or-tools is now available in BCR
https://github.com/google/or-tools/issues/4551

This implicitly also pushed protobuf, abseil-cpp and fuzztest
dependencies, so make that state of versions explicit in the
root project.

Newer protobuf does not compile with clang 17 anymore [^1], so
update clang (this might also help fix some clang-format differences
we see in issue #2006)

protoc-gen-validate did not work with newer protobuf rules anymore,
but luckily it got updated in BCR in the meantime ans works without
local patch now.
Grpc now also exhibit a compile error, so update that to head
as well.

Issues: #931

[^1]: https://github.com/protocolbuffers/protobuf/issues/20645