[XLS] Switch from `--materialize_internal_fifos` to `--fifo_module`
We switch from a hardcoded name for a user-provided FIFO implementation to a user-provided flag. For now, if no module is provided, we still default to `xls_fifo_wrapper`; to disable this, users may pass an empty string for the name (e.g., `--fifo_module=`).

`--materialize_internal_fifos` will still work for now; it is interpreted as changing the default for `--fifo_module` to an empty string, and will error if passed at the same time as a non-empty string is passed to `--fifo_module`.

PiperOrigin-RevId: 739326347