Extracted the optimization of merging consecutive PrioritySelect and OneHotSelect from the select simplification pass. This merging optimization is moved to its own pass called select merging. The related tests that were in select_simplification_pass_test.cc are moved to the new file select_merging_pass_test.cc.
This extraction is needed to defer the merging optimization towards the end of the optimization pipeline as it often blocks other optimizations (e.g., translating a PrioritySelect with a single non-zero case to a mask operation).
This happens because these optimizations target only simple PrioritySelect and OneHotSelect with two (not more) possible inputs.

PiperOrigin-RevId: 740759521