Simplify conditions in else-if blocks: - Reduces XLS runtime - Improves QoR slightly - Prevents some mutual exclusion errors
Previously, when variables were assigned in previous blocks of an if
statement and then referenced in the condition of an else if,
the full select generated in the assignment within the previous block
was used in the condition. This was logically correct but overly complex,
and could confuse the mutual exclusion pass for IO ops within the if-else.

For example:

    if(x == 5) {
        x = 15;
    } else if(x == 15) { // this x will be a select, not just the original x
        x = -2;
    }

This does not apply to switches as the conditional expression is
generated once at the start of translation.

PiperOrigin-RevId: 741172280