Fix package_bazel_build to work with bzlmod.
The runfiles manifest now includes a per-executable file named `_repo_mapping`. This uses the same name in each executable's runfiles but is unique in the cache, e.g. interpreter_main.repo_mapping. It doesn't appear that we need this file, so ignore it.

Also, the manifest for our ir_viz app seems to now include a file called 'Lorem sipsum.txt' from a Python dependency. This is not escaped in any way in the manifest mapping, so we need to just assume that there are at most two fields per line.

PiperOrigin-RevId: 741283128