Remove codegen specialization for priority selects.
We previously specialized the codegen of priority selects when the selector was never zero. The specialization would call `$error()` if the selector was zero and propagate `X`, the idea being that this form might synthesize to something smaller.

Unfortunately, this specialization causes more headaches than it is worth. The `$error()` call isn't a real assertion, so it could spuriously fire. We also generally prefer to avoid propagating X explicitly.

Benchmarks showed that this specialization didn't give a net benefit on our corpus, so it seems better to avoid the headaches by removing it.

Fixes google/xls#1481.

PiperOrigin-RevId: 742428399