Improve precision of the rules for index and verbosity values in TIv2.
Examples like like `a[u8:1]` and `vtrace_fmt!(u32:5, "msg");` work in v1 and the values in the position of the `u8:1` and `u32:5` are typed as the user desires, provided they are acceptable types.

So we cannot unify those types with a static, implied type annotation of u32 or s64, respectively. Instead, this change applies the v1 rules at validation time. For vtrace_fmt!, we collect the rules that are enforced by different downstream phases.

PiperOrigin-RevId: 743219737