Support more verilog-like block simulation semantics.
The current block simulation system had a not very well document feature where it would unconditionally add a synthetic flop across every output port and tap the final stabilized value of that flop to get its output port values. This is often what is actually desired since it avoids the need for all xls blocks to have flopped outputs to simulate but sometimes this is undesirable.

This adds a new `BlockEvaluator::OutputPortSampleTime` argument to continuation creation methods for the block jit/interpreter. This allows one to explicitly control when in the cycle the output-port values are tapped. If passed with `kAtLastPosEdgeClock` the old behavior of acting like there is an implicit flop will be preserved and the output will be read at the instant before the registers flip to new values. If `kAtLastClock` is used then values will be read after everything has stabilized and a full clock cycle has completed.

Note that for `kAfterLastClock` this value will be result of evaluating the block with the cycle's inputs but the registers that are set after the cycle finishes. This is not always a useful piece of information depending on the design being simulated.

Note this is implemented by running the block through simulation twice, once to update registers and get the `AtLastPosEdgeClock` taps and another to get the final tap values.

PiperOrigin-RevId: 744755958