Change v1 tests to use range expr instead of function by default.
Also enable TIv2 for tests that use the expr form.

The plan is to ultimately get rid of the function, which does not behave like a normal function in type checking (i.e. the purported regular params are really parametrics governing the returned array size). So we will just keep one test around in each test suite that proves the function form works until fully deprecated.

PiperOrigin-RevId: 744865990