Introduced the minimum viable product for the resource sharing optimization.
This version of the optimization is only capable to fold a single IR multiplication into another multiplication node that has the same bitwidth of the former for all operands and result.

This new pass is not enabled by default.
To use it, pass the option "--enable_resource_sharing".

This pass includes the necessary analyses to identify the folding opportunities like the mutually-exclusive analysis.

PiperOrigin-RevId: 745283005