First extension to the minimum viable product for the resource sharing optimization.
This extension enables the optimization to fold an arbitrary number of IR multiplications into a single one.

Like the previous version of this optimization, the IR multiplications involved in a folding must have the same bitwidth for their operands as well as for their outputs.

As before, the resource sharing pass is not enabled by default.
To use it, please pass the option "--enable_resource_sharing" when the optimizer is invoked.

PiperOrigin-RevId: 748353389