[xls][mlir] Enable referring to type inside package in linkage.
Previously one couldn't use the types from the package being imported in the template instantiation. Referring to current package requires either knowing the package name or reading in the file and concating with snippet.

But additionally, just using the package name is not sufficient if one wanted to use types not exported by the packaged referenced. An alternative would be adding import statement support. I think at that point this rather becomes plain inline DSLX vs restricted to single function imports.

This follows therefore the latter approach. This does have the downside that one needs to use a name not overlapping with one inside the module - could also do some prefixing etc to reduce the likelihood but the errors returned may get worse.

PiperOrigin-RevId: 750133696