[dslx]: Add **EXPERIMENTAL** annotations for channels.
DSLX's mechanisms for configuring channels are limited. We currently have:

1) Set fifo depth in a channel declaration via `chan<TYPE, u32:DEPTH>`. The values of the rest of the fifo config are based on if depth is 0 or not, which is not very explicit or configurable. This form is  surprising and not very readable (I say this as the person who added it).
2) Set `default_fifo_config` as a flag for IR conversion. This is nice if you have a lot of identical channels you don't want to configure, but is not very granular.

This adds a third mechanism! The feature is hidden behind the module-level `#![feature(channel_attributes)]` and is mutually exclusive with setting the FIFO via option 1). It is weaker than 1) in the sense that it won't constexpr eval the fields, but it is otherwise stronger in every sense.

You configure a channel by writing `#[channel(<key=value, ...>)]`. The keys are the elements of `ChannelConfig` and its contained `FifoConfig`. For implementation convenience, the annotation is on the channel decl, so it looks something like

```
let (a, b) =
#[channel(depth=0)]
chan<u32>("foo");
```

Related: google/xls#1023, google/xls#1561, and google/xls#687.
PiperOrigin-RevId: 750286607