[XLS] Fix use-after-free bug in proc state flattening pass.
If an earlier state element's value was exactly the StateRead of a later state element, we'd end up attempting to create a Next node for it referencing the deleted StateRead, causing use-after-free bugs. To fix this, just as we already do for predicates, we insert an identity to let existing logic patch up the dependency.

PiperOrigin-RevId: 752723193