Fix issue with slicing a destructured loop accumulator element.
There are a couple of issues with a loop body trying to yield a slice of a destructured accumulator element. It doesn't play well with the post-processing of NameDefTree on the converter side, or the somewhat circular flow of type information in loops on the populate side. This change addresses both issues.

PiperOrigin-RevId: 754009194