Change XLS[cc] to use TrackedBValues, in preparation for continuations.
The TrackedBValue class is a wrapper for xls::BValue that enables getting an ordered list of every BValue currently instantiated in XLS[cc].This is deterministic because XLS[cc] generates deterministic IR, and xls::BValues are created in IR generation.

Continuations enable splitting a function into slices, which may be placed
into different proc activations. They expose every "live" value in XLS[cc], which could be referenced after the slice point, at the time the split is created.

The reason it is insufficient to generate continuations for each variable
is that values are kept by XLS[cc] in various places scattered around the code for later use. For example, in an if-else statement, the condition for the if is kept on the stack to be later negated for the else.

PiperOrigin-RevId: 754089654