This change resolves a bug in the resource sharing pass.
The bug occurs when multiple multiplications (called the sources of the folding) can be folded into a multiplication (called the destination of the folding).
In this case, the sources have to be all mutually exclusive between each other, on top of being mutually exclusive with the destination.
This change guarantees the enforcement of these exclusivity relations.

PiperOrigin-RevId: 754151566