Replace `apfloat::upcast` with `apfloat::upcast_daz` (previous behavior) and `apfloat::upcast_with_denorms`.
The latter is useful when we want to convert reduced precision floats to higher precision floats, allowing denormal inputs to turn into a normal output.

Also, update the constraint on the exp/fraction sizes for upcast to match the assumptions in the implementation. The function does not fully cover larger total sizes where one of exp/fraction sizes is reduced. While this is technically a change in behavior (new invalid inputs), previous usage in this scenario could return the wrong result.

PiperOrigin-RevId: 754991280