[opt] Only run XLS IR verification at the start & end of optimization
We previously ran it after each pass that made a change, which (for sufficiently complex blocks) could result in it taking a substantial amount of time.

The previous behavior is now behind the `--debug_optimizations` flag, which also enables lazy query-engine cache validation. The idea is that developers debugging optimizations will want to use this flag, but users can leave it disabled for better performance.

PiperOrigin-RevId: 755650905