Fix the speed of the slow builtin function unit tests in TIv2.
Some of the tests exercise a flaw in unification caching where an InferImplicitParametrics call may repeatedly resolve the same heavy type variable within the scope of one resolution call that has a filter. The general caching logic will refuse to capture the result that was influenced by the filter.

This change makes it so that such globally useless results can be retained within the scope of one overarching resolution request, thereby getting the builtin test suite from ~23 sec to ~3 sec.

PiperOrigin-RevId: 756784291