Avoid repeated work for identical parametric contexts in TIv2.
Generic utilities may be invoked a large number of times in a target, with parametrics that amount to the same values. In such cases there is no point in having separate caching contexts or repeatedly analyzing the callee function, so this change avoids both of those things. This change also eliminates some unnecessary cache invalidation scenarios.

This cuts 80% of the std.x typechecking time.

PiperOrigin-RevId: 756797095