Abstract scheduling graph into separate data structure for the SDC scheduler. The scheduling graph encapsulates all information from the IR graph required for scheduling. The SDC scheduler now holds a ScheduleGraph rather than a FunctionBase*.
This is groundwork for synchronous procs work as it allows graphs other than those which correspond to a single FunctionBase.

Also, as a minor clean up, don't mutate the IR (setting the II on a proc) within the SDC scheduler. Instead set that value after scheduling is complete outside of the scheduler itself.

PiperOrigin-RevId: 758237507