https://github.com/google/xls/issues/1904
Add support for range expression to represent end-inclusive value of a integral type's range, by adding the equal sign before the end expression as in Rust syntax. For example, u8:0..=u8:255 represents the array of integers in [0..255] rather than [0..255). Previously there is no easy way to access uN[x]::MAX in a range. This also applies to ranges in match arm.

PiperOrigin-RevId: 758898841