Optimize IntervalSet::Combine
This is a relatively hot-spot in cond-spec and other passes. This replaces the old implementation with one which manually iterates through the intervals and avoids allocating as much as possible. This is up to ~6x faster than the naive implementation (on a somewhat contrived test cases).

Actual use cases show improvement of ~5% overall opt perf on some targets.

PiperOrigin-RevId: 759224647