Fix function signature in apfloat that incorrectly compiles with TIv1.
In a case like:

```
struct S<N: u32> {
  n: uN[N]
}

fn factory<N: u32>() -> S {
  S { n: uN[N]:0 }
}

const X = factory<u32:8>();
```

v1 will magically pretend the formal return type is S<N>. This appears to be a bug in v1 (mismanagement of parametrics by name), because if we change the name of the parametric on the actual struct def to a different name than the function parametric, it fails.

Supporting this intentionally in v2 would be a pain, and probably lead to users writing more obscure DSLX code, so let's consider it a bug and fix the example in apfloat.

PiperOrigin-RevId: 761541826