[XLS] Add a more-precise model of OneHot for IntervalSets
Rather than converting our IntervalSets to ternary values, trying to model the impact of `one_hot` there, then converting back with poor resolution - we instead test if each possible result bit can be set, and add it to our resulting IntervalSet if it can.

NOTE: The runtime performance of this should be `O(I*n^2)`, where `I` is the  number of intervals in the input `val`, and `n` is `val.BitCount()`, since `CoversTernary` runs in `O(I*n)` time.
PiperOrigin-RevId: 761588016