[XLS] When comparing TreeBitLocations, compare node ID rather than pointer
Do the same when caching BDD trees for each node in the `node_locations_` cache in the BDD query engine.

This is generally both a correctness & a safety improvement. In this specific case, it was found as a fix for nondeterminism! Due to false matches when memory locations were reused for new Nodes, the BDD query engine's variable caches could nondeterministically match with an (no-longer-used) variable created for a deleted Node. This was safe, but resulted in effective changes to the variable order, which could cause the BDD to saturate in different places nondeterministically.

PiperOrigin-RevId: 761693685