Avoid large alloca's in jitted code.
If one attempted to jit an IR with many exceptionally large values (especially using them in a counted-for loop) it was possible for the generated code to overflow the stack quite easily. To avoid this the JIT is changed to instead allocate large function-local buffers on the heap using callbacks.

NB The underlying change to map that causes the explosion in stack size also causes unopt ir-interpreter to be a lot slower which is the cause of https://github.com/google/xls/issues/2263.

Issue: https://github.com/google/xls/issues/2263
Fixes: https://github.com/google/xls/issues/2049
PiperOrigin-RevId: 765321570