Fix std::aligned_alloc call for asan
In asan mode the std::aligned_alloc will fail if the size is not a multiple of the alignment. This is technically allowed by the c-std (which just says that the allocation may fail) however most allocators will support any pair of size and alignment as long as both are relatively normal. The default ASAN allocator on the other hand follows the older more restrictive c11 spec which requires that calls with a size which is not a multiple of alignment to fail. By overallocating we support the more conservative behavior of ASAN.

PiperOrigin-RevId: 766270339