Added a reassociation IR optimization that favors the pairing of equal variables in an addition expression if the variables are the same.
To double a number in hardware design, it is more efficient to multiply the number by 2 than to add the number to itself. This is because we can perform a left bit shift. This optimization pairs variables that are the same so that a lower-level compiler optimization down the line can convert, for example, (a + a) to (a << 1).

Example:
(a + b + a + b + a) -> (((a + a) + (b + b)) + a)
PiperOrigin-RevId: 767178332