Fix bug where TIv2 associates TypeInfo for helper of imported function with empty caller env.
This was affecting the IR conversion of a number of targets, which would complain of a null `TypeInfo` pointer.

The issue was that the mapping of ParametricContext to ParametricEnv was in the converter object and therefore siloed by module, so the converter could not look up a ParametricEnv for a foreign caller. This change moves the remaining ParametricContext-related data to the corpus-wide table.

PiperOrigin-RevId: 769159531