[xls] Fix narrowing pass dynamic-slice bounds error
When we can infer a value for the slice start that doesn't fit in an int64_t, we previously ended up violating the check. With this fix, we instead recognize that this means the slice starts after the end of any allowed value, and replace the slice with a literal zero.

PiperOrigin-RevId: 769182009