Began implementation of the IR Fuzzer which currently supports single-width adder trees.
The IR Fuzzer is a new fuzzer that seeks to randomly generate IR from a struct domain (such as protobufs). It uses gtest tools, making it more compatible with Google infrastructure. It is different from the existing fuzzer because it uses gtest and does not start with DSLX.

The IR Fuzzer currently supports the random generation of single-width adder trees. It does this by accepting a randomly generated FuzzProgram protobuf into the IrFuzzBuilder class, processing a list of FuzzOperations from the FuzzProgram, and converting the operations into IR operations that are placed inside a stack. A single IR node value is returned, representing the randomly generated IR.

PiperOrigin-RevId: 770793936