Added const eval of if-expr's conditional in parametric function to eliminate branch that is never taken.
Since in parametric function we often branch on the parametric arguments to do different things, while the untaken branch may contain statements end up with type errors if they are type checked, which is superfluous. If we can prove the conditional is a const expression, only the taken branch will be type-checked and compiled.

PiperOrigin-RevId: 772295944