Make sure the llvm-build does not pick up the system zlib header.
The includes of zlib in the llvm-project uses angle brackets
instead of double-quotes implying to pick the system header
instead of the one provided by the BUILD context.
If the zlib header is not installed on the system, this resulted
in a build-breakage for me.

Add a patch to the llvm toolchain to fix the files using the
wrong quotes.