Scheduling optimization.
In practice, XLS IR often has a large number of nodes with 0 delay. Some common cases are if the node is a literal, tuple_index, or concat. There are many cases where it is known that a node will schedule into the same stage as a user or operand, but before now we still created a set of variables and constraints and let the solver figure it out.

This optimization recognizes some simple cases and avoids creating more variables and constraints for them. The idea is to reduce the scale of the problem we throw at the solver to improve runtime for large designs.

PiperOrigin-RevId: 775401213