modules: zstd: Add ram passthrough proc and wrappers for muxes/demuxes
This commit adds a passthrough proc which is attached to the output of
the RamDemux proc. Because of this, RAM rewriting is used on the outputs
of the passthrough instead of the outputs of the RamDemux.

This resolves the problem spotted in the Verilog simulation,
where artificial responses are sent from RamDemux despite
no request being made previously.

Signed-off-by: Robert Winkler <rwinkler@antmicro.com>