[xls][mlir} Fix bug `-array-to-bits` related to `xls.dynamic_bit_slice`.
This CL fixes a bug that occurred in the `-array-to-bits` pass during the legalization of `xls.array_slice`, which produced an incorrect `start` value for the resulting `xls.dynamic_bit_slice` op. The wrong computation comes from the fact that XLS uses left-to-right offsets for `array`s, including the original `xls.array_slice` op, but right-to-left offsets for `bits`, including the created `xls.dynamic_bit_slice`. The previous legalization pattern ignored the direction change during rewriting, so the fix consists of converting the left-to-right offset into a right-to-left offset.

The CL also adds missing punctuation in an unrelated file and adds the `ArithDialect` to the dependencies of the `-array-to-bits` pass because this and other patterns may create ops from that dialect.

PiperOrigin-RevId: 777914087