Generate a list of all passes in markdown.
This uses the pass registry and a clang-libtool to generate documentation for every pass accessible by the pipeline proto or the --passes argument in a markdown file. The documentation consists of the 'short_name' used to specify the pass, the long-name and the comment associated with the pass definition.

Compound passes defined by the pipeline proto have a new field for any comments.

Unfortunately due to differences in how libcxx is included in the build this does not currently work in OSS (because there clang uses its own configured libcxx which our libtool version can't see and might not be the correct version of). This should be runnable if you provide a manual libcxx of the right commit.

Bug: https://github.com/google/xls/issues/2556
PiperOrigin-RevId: 780644038